Keep graph decoder tests stable under mixed precision
The CUDA graph-decoder path can run the MLP under autocast, yielding half-precision edge probabilities. The reconstructed adjacency tensor must use the same dtype before indexed assignment.

Constraint: pytest runs the graph decoder on CUDA when torch_geometric is installed
Confidence: high
Scope-risk: narrow
Directive: Preserve decoder output dtype alignment when changing AMP or graph reconstruction code
Tested: pytest tests/test_models.py::TestGraphEncoder::test_graph_decoder -q
Tested: pytest -q
Not-tested: CPU-only torch_geometric environment

+++ gahib/core/graph_utils.py
@@ -129,7 +129,11 @@ class MLPDecoder(nn.Module):
         row, col = edge_index
         edge_features = torch.cat([z[row], z[col]], dim=1)
         edge_probs = self.mlp(edge_features).squeeze()
-        adj_recon = torch.zeros((num_nodes, num_nodes), device=z.device)
+        adj_recon = torch.zeros(
+            (num_nodes, num_nodes),
+            device=z.device,
+            dtype=edge_probs.dtype,
+        )
         adj_recon[row, col] = edge_probs
         return adj_recon
